examples/resources/requirements.txt
jupyter
openassetio>=v1.0.0b1rev0
openassetio-manager-bal>=v1.0.0a13
openassetio-mediacreation
Pillow
